week-01/build_your_own_infrastructure/gcp/provision/terraform/modules/cloud_storage/functions/data_ingestion/requirements.txt
requests==2.25.1
google-cloud-storage